+++ core/src/main/java/nl/nn/adapterframework/receivers/SambaListener.java
@@ -1,5 +1,5 @@
 package nl.nn.adapterframework.receivers;
 
-public class SambaListener extends Samba2Listener {
+public class SambaListener {
 
 }