+++ core/src/main/java/nl/nn/adapterframework/util/Misc.java
@@ -568,6 +568,9 @@ public class Misc {
 		try {
             return (String) Class.forName("nl.nn.adapterframework.util.IbmMisc").getMethod("getApplicationDeploymentDescriptorPath").invoke(null);
 		} catch (Exception e) {
+			if("WAS".equals(AppConstants.getInstance().getString("application.server.type", ""))) {
+				throw new IOException(e);
+			}
             log.debug("Caught NoClassDefFoundError for getApplicationDeploymentDescriptorPath, just not on Websphere Application Server: " + e.getMessage());
             return null;
         }